+++ CHANGELOG.md
@@ -4,13 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-### 1.7.0  - (Unreleased)
+### 2.0.0  - (Unreleased)
 
 Most of the app has be rewritten in this update
 
-#### Changed
+#### Added
 
 - Animated bottom bar
+- Meditation
+- Lots more languages
+
+#### Changed
+
+- New bottom sheets
+- Lists moved to main page
+- Lots of optimisations and fixes
 
 ### 1.6.0 - (2020-12-03)